implement tag creation and linking

+++ lib/features/file_explorer/data/repository/file_tag_link_repository_impl.dart
@@ -1,3 +1,8 @@
+import 'package:fpdart/fpdart.dart';
+import 'package:tagsurf_flutter/features/file_explorer/core/error/exceptions.dart';
+import 'package:tagsurf_flutter/features/file_explorer/core/error/failure.dart';
+import 'package:tagsurf_flutter/features/file_explorer/core/error/file_tag_links_failures.dart';
+import 'package:tagsurf_flutter/features/file_explorer/core/error/general_failures.dart';
 import 'package:tagsurf_flutter/features/file_explorer/data/data_sources/database/app_database.dart';
 import 'package:tagsurf_flutter/features/file_explorer/data/models/file_tag_link.dart';
 import 'package:tagsurf_flutter/features/file_explorer/domain/entities/file_entity.dart';
@@ -10,40 +15,77 @@ class FileTagLinkRepositoryImpl implements FileTagLinkRepository {
   FileTagLinkRepositoryImpl(this._appDatabase);
 
   @override
-  Future<List<FileEntity>> getFilesByTag({required TagEntity tag}) async {
-    final filesModels =
-        await _appDatabase.fileTagLinkDao.getFilesByTagName(tag.name);
-    return filesModels
-        .map((fileModel) => FileEntity.fromModel(fileModel))
-        .toList();
+  Future<Either<Failure, List<FileEntity>>> getFilesByTag(
+      {required TagEntity tag}) async {
+    try {
+      final filesModels =
+          await _appDatabase.fileTagLinkDao.getFilesByTagName(tag.name);
+      return Right(filesModels
+          .map((fileModel) => FileEntity.fromModel(fileModel))
+          .toList());
+    } on DatabaseException {
+      return Left(DatabaseFailure());
+    }
   }
 
   @override
-  Future<List<TagEntity>> getTagsByFile({required FileEntity file}) async {
-    final tagsModels =
-        await _appDatabase.fileTagLinkDao.getTagsByFilePath(file.path);
-    return tagsModels.map((tagModel) => TagEntity.fromModel(tagModel)).toList();
+  Future<Either<Failure, List<TagEntity>>> getTagsByFile(
+      {required FileEntity file}) async {
+    try {
+      final tagsModels =
+          await _appDatabase.fileTagLinkDao.getTagsByFilePath(file.path);
+      return Right(
+          tagsModels.map((tagModel) => TagEntity.fromModel(tagModel)).toList());
+    } on DatabaseException {
+      return Left(DatabaseFailure());
+    }
   }
 
   @override
-  Future<void> linkFileAndTag(
+  Future<Either<Failure, void>> linkFileAndTag(
       {required String filePath, required String tagName}) async {
-    _appDatabase.fileTagLinkDao.insertFileTagLink(
-        FileTagLinkModel(filePath: filePath, tagName: tagName));
+    try {
+      final existingLink =
+          await _appDatabase.fileTagLinkDao.getFileTagLink(filePath, tagName);
+      if (existingLink == null) {
+        await _appDatabase.fileTagLinkDao.insertFileTagLink(
+            FileTagLinkModel(filePath: filePath, tagName: tagName));
+        return const Right(null);
+      } else {
+        return Left(LinkDuplicateFailure(file: filePath, tag: tagName));
+      }
+    } on DatabaseException {
+      return Left(DatabaseFailure());
+    }
   }
 
   @override
-  Future<void> unlinkFileAndTag(
+  Future<Either<Failure, void>> unlinkFileAndTag(
       {required String filePath, required String tagName}) async {
-    _appDatabase.fileTagLinkDao.deleteFileTagLink(
-        FileTagLinkModel(filePath: filePath, tagName: tagName));
+    try {
+      final existingLink =
+          await _appDatabase.fileTagLinkDao.getFileTagLink(filePath, tagName);
+      if (existingLink != null) {
+        await _appDatabase.fileTagLinkDao.deleteFileTagLink(
+            FileTagLinkModel(filePath: filePath, tagName: tagName));
+        return const Right(null);
+      } else {
+        return Left(LinkNotExistFailure(file: filePath, tag: tagName));
+      }
+    } on DatabaseException {
+      return Left(DatabaseFailure());
+    }
   }
 
   @override
-  Future<List<FileEntity>> getUntaggedFiles() async {
-    final filesModels = await _appDatabase.fileTagLinkDao.getUntaggedFiles();
-    return filesModels
-        .map((fileModel) => FileEntity.fromModel(fileModel))
-        .toList();
+  Future<Either<Failure, List<FileEntity>>> getUntaggedFiles() async {
+    try {
+      final filesModels = await _appDatabase.fileTagLinkDao.getUntaggedFiles();
+      return Right(filesModels
+          .map((fileModel) => FileEntity.fromModel(fileModel))
+          .toList());
+    } on DatabaseException {
+      return Left(DatabaseFailure());
+    }
   }
 }
